+++ src/main/java/cl/duoc/ms_tiendas/dto/TiendaResponse.java
@@ -1,6 +1,5 @@
 package cl.duoc.ms_tiendas.dto;
 
-
 import cl.duoc.ms_tiendas.model.EstadoTienda;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -13,7 +12,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class TiendaResponse {
 
-    private Integer   id;
+    private Integer id;
     private String nombre;
     private String descripcion;
     private String telefono;
@@ -24,7 +23,7 @@ public class TiendaResponse {
     private EstadoTienda estado;
 
     // ID del dueno (para que el frontend sepa si el usuario actual es dueno)
-    private Long idUsuarioDueno;
+    private Integer idUsuarioDueno;
 
     // Nombre del dueño enriquecido desde ms-login (puede ser null si ms-login no responde)
     private String nombreDueno;
